requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.16.0
peft>=0.7.0
bitsandbytes>=0.41.0

# ML Foundry API
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.0
jsonlines>=4.0.0

# Evaluation & Metrics
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
wandb>=0.16.0

# Code Analysis
ast-monitor>=0.1.0
tokenize-rt>=5.2.0

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0